Clip_Service/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
transformers==4.55.4
torch==2.8.0
Pillow==11.3.0